Lockscreen charging info: show decimal in battery temperature

+++ packages/SettingsLib/src/com/android/settingslib/fuelgauge/BatteryStatus.java
@@ -53,12 +53,12 @@ public class BatteryStatus {
     public final int maxChargingCurrent;
     public final int maxChargingVoltage;
     public final int maxChargingWattage;
-    public final int temperature;
+    public final float temperature;
     public final boolean oemFastChargeStatus;
 
     public BatteryStatus(int status, int level, int plugged, int health,
             int maxChargingCurrent, int maxChargingVoltage,
-            int maxChargingWattage, int temperature, boolean oemFastChargeStatus) {
+            int maxChargingWattage, float temperature, boolean oemFastChargeStatus) {
         this.status = status;
         this.level = level;
         this.plugged = plugged;
